Switch to a version number permitted by Dist::Zilla::Plugin::BumpVersionAfterRelease
According to http://search.cpan.org/~dagolden/Dist-Zilla-Plugin-
BumpVersionAfterRelease-0.015/lib/Dist/Zilla/Plugin/BumpVersionAfterRelease.pm
which is in use in because of [RewriteVersion] in dist.ini, version
numbers should not be in tuple format.

This change allows dzil build to succeed. I'm assuming that this would
be the next release number and that it can stay like this until
released.